Add files via upload
design of Biometric Voting Machine Using Fingerprint Scanner and 
Arduino for voting. Indian constitution empowers its citizen to 
exercise right to vote. Election decides the future of country, so 
that the system used for voting should be trustworthy. The 
conventional system for voting is Electronic voting machine, has 
many flaws and trust issues. To eradicate malpractice and 
defrauding of the above methods of voting, we have designed an 
advanced system by using arduino and Fingerprint module.